requirements/test.txt
-r production.txt

pytest==3.4.0
pytest-django==3.1.2
pytest-pspec==0.0.3
